refactoring: Int64 -> Int
There is really no point in being anal about the 64-bitness of
integers here, in t. Switching makes interfacing with other
libraries nicer as well.

+++ src/T/Stdlib/Fun.hs
@@ -35,10 +35,10 @@ functions =
   [ Fun "empty?" nullB
   , Fun "length" lengthB
 
-  , Fun "floor" (flip embed0 (floor @Double @Int64))
-  , Fun "ceiling" (flip embed0 (ceiling @Double @Int64))
-  , Fun "round" (flip embed0 (round @Double @Int64))
-  , Fun "int->double" (flip embed0 (fromIntegral @Int64 @Double))
+  , Fun "floor" (flip embed0 (floor @Double @Int))
+  , Fun "ceiling" (flip embed0 (ceiling @Double @Int))
+  , Fun "round" (flip embed0 (round @Double @Int))
+  , Fun "int->double" (flip embed0 (fromIntegral @Int @Double))
 
   , Fun "upper-case" (flip embed0 Text.toUpper)
   , Fun "lower-case" (flip embed0 Text.toLower)
